Create OOP Class
This Python code defines a class called Employee that represents employees in a company. Each Employee instance is assigned a unique ID when it is created. Here's a step-by-step description of what the code does:

The Employee class is defined with a class variable new_id initialized to 1. This variable keeps track of the next available ID for an employee.

The __init__ method is a constructor that gets called automatically when a new Employee instance is created. It assigns the current value of Employee.new_id to the id attribute of the instance being created and then increments Employee.new_id by 1 to prepare for the next instance.

The say_id method prints out the ID of the Employee instance on which it is called.

Two Employee instances (e1 and e2) are created.

When e1 is created, its id is set to 1 because Employee.new_id is 1. Then, Employee.new_id is incremented to 2.

When e2 is created, its id is set to 2 because Employee.new_id is now 2. Then, Employee.new_id is incremented to 3.

Finally, the IDs of both Employee instances (e1 and e2) are printed using the say_id method.